fix: Edusign - init groupes + etudiants
