+++ src/Classes/EduSign/Api/ApiCours.php
@@ -1,0 +1,163 @@
+<?php
+
+namespace App\Classes\EduSign\Api;
+
+use App\Classes\EduSign\DTO\EduSignCourse;
+use App\Classes\EduSign\GetCleApi;
+use App\Repository\EdtCelcatRepository;
+use App\Repository\EdtPlanningRepository;
+use Exception;
+use Symfony\Component\HttpClient\HttpClient;
+
+class ApiCours
+{
+    public function __construct(
+        protected GetCleApi             $getCleApi,
+        protected EdtPlanningRepository $edtPlanningRepository,
+        protected EdtCelcatRepository   $edtCelcatRepository,
+
+    )
+    {
+    }
+
+    public function addCourse(EduSignCourse $course, string $cleApi): string
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('POST', 'https://ext.edusign.fr/v1/course', [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ],
+            'json' => ['course' => $course->toArray()],
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+
+        $data = json_decode($content, true);
+        // accéder à la valeur de l'ID
+        $id = $data['result']['ID'] ?? "";
+
+        if (str_contains($course->type_edt, 'intranet')) {
+            $edt = $this->edtPlanningRepository->findOneBy(['id' => $course->api_id]);
+            $rep = $this->edtPlanningRepository;
+        } elseif (str_contains($course->type_edt, 'celcat')) {
+            $edt = $this->edtCelcatRepository->findOneBy(['id' => $course->api_id]);
+            $rep = $this->edtCelcatRepository;
+        } else {
+            $edt = null;
+            $rep = null;
+        }
+
+        if (null === $edt) {
+            throw new Exception('Course not found for ' . $edt->api_id);
+        }
+
+        if ($edt->getIdEduSign() == null && $rep !== null) {
+            $edt->setIdEduSign($id);
+            $rep->save($edt);
+        }
+
+        return $content;
+    }
+
+    public function updateCourse(EduSignCourse $course, string $cleApi): string
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('PATCH', 'https://ext.edusign.fr/v1/professor/', [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ],
+            'json' => ['course' => $course->toArray()],
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+
+        return $content;
+    }
+
+    public function getCourseIdByApiId(string $apiId, string $cleApi): mixed
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('GET', 'https://ext.edusign.fr/v1/course/get-id/' . $apiId, [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ]
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+        // convertit JSON en tableau associatif PHP
+        $data = json_decode($content, true);
+
+        return $data['result'] ?? "";
+    }
+
+    public function getCourses(?string $id, string $cleApi): mixed
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('GET', 'https://ext.edusign.fr/v1/course/' . $id, [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ]
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+        // convertit JSON en tableau associatif PHP
+        $data = json_decode($content, true);
+
+        return $data['result'] ?? "";
+    }
+
+    public function getAllCourses(string $cleApi): mixed
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('GET', 'https://ext.edusign.fr/v1/course', [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ]
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+        // convertit JSON en tableau associatif PHP
+        $data = json_decode($content, true);
+
+        return $data['result'] ?? null;
+    }
+
+    public function deleteCourse(?EduSignCourse $course, string $cleApi): string
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('DELETE', 'https://ext.edusign.fr/v1/course/' . $course->id, [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ],
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+
+        $edt = $this->edtPlanningRepository->findOneBy(['idEduSign' => $course]);
+        if ($edt) {
+            $edt->setIdEduSign(null);
+            $this->edtPlanningRepository->save($edt);
+        }
+
+        return $content;
+
+    }
+}
