+++ src/Classes/EduSign/Api/ApiGroupe.php
@@ -1,0 +1,106 @@
+<?php
+
+namespace App\Classes\EduSign\Api;
+
+use App\Classes\EduSign\DTO\EduSignGroupe;
+use App\Classes\EduSign\GetCleApi;
+use App\Entity\Groupe;
+use App\Entity\Semestre;
+use App\Repository\GroupeRepository;
+use App\Repository\SemestreRepository;
+use Symfony\Component\HttpClient\HttpClient;
+
+class ApiGroupe
+{
+    public function __construct(
+        protected GetCleApi          $getCleApi,
+        protected GroupeRepository   $groupeRepository,
+        protected SemestreRepository $semestreRepository
+    )
+    {
+    }
+
+    public function getAllGroups(string $cleApi): mixed
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('GET', 'https://ext.edusign.fr/v1/group', [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ]
+        ]);
+
+        $content = $response->getContent();
+        // convertit JSON en tableau associatif PHP
+        $data = json_decode($content, true);
+
+        return $data['result'] ?? "";
+    }
+
+    public function addGroupe(EduSignGroupe $groupe, string $cleApi, ?string $type): mixed
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('POST', 'https://ext.edusign.fr/v1/group', [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ],
+            'json' => ['group' => $groupe->toArray()],
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+
+        $data = json_decode($content, true);
+        // accéder à la valeur de l'ID
+        $id = $data['result']['ID'] ?? "";
+
+        if ($type === 'semestre') {
+            $semestre = $this->semestreRepository->findOneBy(['id' => $groupe->api_id]);
+            if ($semestre && $semestre instanceof Semestre) {
+                if ($semestre->getIdEduSign() === null) {
+                    $semestre->setIdEduSign($id);
+                    $this->semestreRepository->save($semestre);
+                }
+            }
+        } elseif ($type === 'groupe') {
+            $groupeAdd = $this->groupeRepository->findOneBy(['id' => $groupe->api_id]);
+            if ($groupeAdd && $groupeAdd instanceof Groupe) {
+                if ($groupeAdd && $groupeAdd->getIdEduSign() === null) {
+                    $groupeAdd->setIdEduSign($id);
+                    $this->groupeRepository->save($groupeAdd);
+                }
+            }
+        }
+
+        return $content;
+    }
+
+    public function deleteGroupe(EduSignGroupe $groupe, string $cleApi): void
+    {
+        $client = HttpClient::create();
+
+        $response = $client->request('DELETE', 'https://ext.edusign.fr/v1/group/' . $groupe->id_edu_sign, [
+            'headers' => [
+                'Content-Type' => 'application/json',
+                'Authorization' => 'Bearer ' . $this->getCleApi->getCleApi($cleApi),
+            ],
+        ]);
+
+        $statusCode = $response->getStatusCode();
+        $content = $response->getContent();
+
+        $groupe = $this->groupeRepository->findOneBy(['idEduSign' => $groupe->id_edu_sign]) ?? $this->semestreRepository->findOneBy(['idEduSign' => $groupe->id_edu_sign]);
+        if ($groupe) {
+            $groupe->setIdEduSign(null);
+            if ($groupe instanceof Groupe) {
+                $this->groupeRepository->save($groupe);
+            } elseif ($groupe instanceof Semestre) {
+                $this->semestreRepository->save($groupe);
+            }
+        }
+
+    }
+}
